Updates after code review
* Renamed default serializer to make its purpose more clear
* Updated owner attribute to use entities relationships
* Added dependency on lodash from backstage library
* Fixed testing mocking with lodash cloneDeep
* Added dependency on @backstage/plugin-techdocs-common
* Updated techdocs annotation to use library constant
