+++ plugins/datadog-entity-sync-node/package.json
@@ -28,14 +28,20 @@
     "@backstage/catalog-model": "^1.7.3",
     "@backstage/plugin-catalog-node": "^1.16.3",
     "@backstage/plugin-events-node": "^0.4.10",
+    "@backstage/plugin-techdocs-common": "^0.1.0",
     "@backstage/types": "^1.2.1",
     "@datadog/datadog-api-client": "^1.34.1",
     "git-url-parse": "^16.1.0"
   },
+  "peerDependencies": {
+    "lodash": "*"
+  },
   "devDependencies": {
     "@backstage/backend-test-utils": "^1.4.0",
     "@backstage/cli": "^0.32.0",
-    "@types/git-url-parse": "^16.0.2"
+    "@types/git-url-parse": "^16.0.2",
+    "@types/lodash": "^4",
+    "lodash": "^4.17.21"
   },
   "files": [
     "dist",
